Add files via upload
This GitHub repository contains a web scraping project that utilizes Python and the BeautifulSoup library to extract data from the website "Quotes to Scrape". The project scrapes the author name, quote, and associated tags for each quote on the website. The data is then cleaned and organized using pandas, allowing for insightful data analysis.